test: split SWM snapshot and readiness scenarios (#2106)
